ZaloApi Function

+++ app/src/main/java/com/example/foodapp/Domain/Foods.java
@@ -8,7 +8,7 @@ public class Foods implements Serializable {
     private boolean BestFood;
     private int Id;
     private int LocationId;
-    private double Price;
+    private long Price;
     private String ImagePath;
     private int PriceId;
     private double Star;
@@ -50,11 +50,11 @@ public class Foods implements Serializable {
         LocationId = locationId;
     }
 
-    public double getPrice() {
+    public long getPrice() {
         return Price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(long price) {
         Price = price;
     }
